// src/interfaces/pokemonInterface.tsx
export interface PokemonSimple {
    id: string;
    name: string;
    picture: string;
    color?: string;
}